+++ src/FolkerKinzel.CsvTools.Mappings.Tests/CsvConverterExtensionTests.cs
@@ -47,7 +47,7 @@ public class CsvConverterExtensionTests
 
         CsvMapping mapping = CsvMappingBuilder.Create().AddProperty("A", 0, converter).AddProperty("B", 1, converter).Build();
 
-        string csv = values.ToCsv(2, new ToCsvIntl<(int A, int B)>( mapping, (tuple, dyn) => { dyn.A = tuple.A; dyn.B = tuple.B; }));
+        string csv = values.ToCsv(2, new CsvFromIntl<(int A, int B)>( mapping, (tuple, dyn) => { dyn.A = tuple.A; dyn.B = tuple.B; }));
 
         (int A, int B)[] results = CsvConverter.Parse<(int A, int B)>(csv, mapping, dyn => (dyn.A, dyn.B), isHeaderPresent: false);
 
@@ -73,7 +73,7 @@ public class CsvConverterExtensionTests
     public void ToCsvTest5()
     {
         CsvMapping mapping = CsvMappingBuilder.Create().AddProperty("A", StringConverter.CreateNullable()).Build();
-        string s = new string[] { "Hi" }.ToCsv(new ToCsvIntl<string>(mapping, (string data, dynamic mapping) => mapping.A = data));
+        string s = new string[] { "Hi" }.ToCsv(new CsvFromIntl<string>(mapping, (string data, dynamic mapping) => mapping.A = data));
         Assert.AreEqual("""
             A
             Hi
@@ -109,7 +109,7 @@ public class CsvConverterExtensionTests
 
         CsvMapping mapping = CsvMappingBuilder.Create().AddProperty("A", converter).AddProperty("B", converter).Build();
 
-        values.SaveCsv(filePath, new ToCsvIntl<(int A, int B)>( mapping, (tuple, dyn) => { dyn.A = tuple.A; dyn.B = tuple.B; }));
+        values.SaveCsv(filePath, new CsvFromIntl<(int A, int B)>( mapping, (tuple, dyn) => { dyn.A = tuple.A; dyn.B = tuple.B; }));
 
         using CsvReader<(int A, int B)> reader = CsvConverter.OpenRead<(int A, int B)>(filePath, mapping, dyn => (dyn.A, dyn.B));
 
@@ -146,7 +146,7 @@ public class CsvConverterExtensionTests
 
         CsvMapping mapping = CsvMappingBuilder.Create().AddProperty("A", 0, converter).AddProperty("B", 1, converter).Build();
 
-        values.SaveCsv(filePath, 2, new ToCsvIntl<(int A, int B)>( mapping, (tuple, dyn) => { dyn.A = tuple.A; dyn.B = tuple.B; }));
+        values.SaveCsv(filePath, 2, new CsvFromIntl<(int A, int B)>( mapping, (tuple, dyn) => { dyn.A = tuple.A; dyn.B = tuple.B; }));
 
         using CsvReader<(int A, int B)> reader = CsvConverter.OpenRead<(int A, int B)>(filePath, mapping, dyn => (dyn.A, dyn.B), isHeaderPresent: false);
         (int A, int B)[] results = [.. reader];