+++ src/FolkerKinzel.CsvTools.Mappings/CsvWriter_TSource.cs
@@ -5,14 +5,14 @@ namespace FolkerKinzel.CsvTools.Mappings;
 
 /// <summary>Writes data of any type as CSV (RFC 4180).</summary>
 /// 
-/// <typeparam name="TData">
+/// <typeparam name="TSource">
 /// Generic type parameter for the data type that the 
-/// <see cref="CsvWriter{TData}"/> can write as CSV row.
+/// <see cref="CsvWriter{TSource}"/> can write as CSV row.
 /// </typeparam>
-public sealed class CsvWriter<TData> : IDisposable
+public sealed class CsvWriter<TSource> : IDisposable
 {
     private readonly CsvWriter _writer;
-    private readonly ToCsv<TData> _converter;
+    private readonly CsvFrom<TSource> _converter;
     private bool _disposed;
 
     /// <summary>
@@ -20,15 +20,15 @@ public sealed class CsvWriter<TData> : IDisposable
     /// </summary>
     /// <param name="writer">A <see cref="CsvWriter"/> instance.</param>
     /// <param name="mapping">The <see cref="CsvMapping"/> used to 
-    /// convert <typeparamref name="TData"/> to CSV.</param>
+    /// convert <typeparamref name="TSource"/> to CSV.</param>
     /// <param name="conversion">
     /// <para>
-    /// A method that fills the content of a <typeparamref name="TData"/>
+    /// A method that fills the content of a <typeparamref name="TSource"/>
     /// instance into the properties of <paramref name="mapping"/>. 
     /// </para>
     /// <para>
-    /// <paramref name="conversion"/> is called with each call to <see cref="Write(TData)"/>
-    /// and it gets the <typeparamref name="TData"/> instance and <paramref name="mapping"/>
+    /// <paramref name="conversion"/> is called with each call to <see cref="Write(TSource)"/>
+    /// and it gets the <typeparamref name="TSource"/> instance and <paramref name="mapping"/>
     /// as arguments. The <see cref="CsvMapping"/> is passed to the method as <c>dynamic</c>
     /// argument: Inside the method the registered <see cref="DynamicProperty"/> instances 
     /// can be used like regular .NET properties, but without IntelliSense ("late binding").
@@ -43,12 +43,12 @@ public sealed class CsvWriter<TData> : IDisposable
     /// <paramref name="mapping"/>, or <paramref name="conversion"/> is <c>null</c>.</exception>
     public CsvWriter(CsvWriter writer,
                      CsvMapping mapping,
-                     Action<TData, dynamic> conversion)
+                     Action<TSource, dynamic> conversion)
     {
         _ArgumentNullException.ThrowIfNull(writer, nameof(writer));
 
         _writer = writer;
-        _converter = new ToCsvIntl<TData>(mapping, conversion);
+        _converter = new CsvFromIntl<TSource>(mapping, conversion);
         _converter.Mapping.Record = _writer.Record;
     }
 
@@ -58,14 +58,14 @@ public sealed class CsvWriter<TData> : IDisposable
     /// <param name="writer">A <see cref="CsvWriter"/> instance.</param>
     /// 
     /// <param name="converter">
-    /// An object that converts a <typeparamref name="TData"/> instance to a 
+    /// An object that converts a <typeparamref name="TSource"/> instance to a 
     /// CSV row.
     /// </param>
     /// 
     /// <exception cref="ArgumentNullException"><paramref name="writer"/>, or 
     /// <paramref name="converter"/> is <c>null</c>.</exception>
     public CsvWriter(CsvWriter writer,
-                     ToCsv<TData> converter)
+                     CsvFrom<TSource> converter)
     {
         _ArgumentNullException.ThrowIfNull(writer, nameof(writer));
         _ArgumentNullException.ThrowIfNull(converter, nameof(converter));
@@ -78,11 +78,11 @@ public sealed class CsvWriter<TData> : IDisposable
     /// <summary>
     /// Writes <paramref name="data"/> as a new CSV row.
     /// </summary>
-    /// <param name="data">The <typeparamref name="TData"/> instance to be written or 
+    /// <param name="data">The <typeparamref name="TSource"/> instance to be written or 
     /// <c>null</c>.</param>
     /// <exception cref="IOException">I/O error.</exception>
     /// <exception cref="ObjectDisposedException">The file was already closed.</exception>
-    public void Write(TData? data)
+    public void Write(TSource? data)
     {
         if (data is not null)
         {